backend/requirements.txt
Django>=4.0
python-decouple
djangorestframework
celery
redis
requests